refactoring utils module name
