+++ sklarpy/__init__.py
@@ -1,2 +1,2 @@
-from sklarpy._utils._serialize import load
-from sklarpy._utils._object_printing import print_full
+from sklarpy.utils._serialize import load
+from sklarpy.utils._object_printing import print_full
