+++ sklarpy/copulas/_distributions/_gaussian_kde.py
@@ -6,7 +6,7 @@ from typing import Union
 from sklarpy.copulas._prefit_dists import PreFitCopula
 from sklarpy.copulas._fitted_dists import FittedCopula
 from sklarpy.copulas import MarginalFitter
-from sklarpy._utils import Params
+from sklarpy.utils._params import Params
 
 __all__ = ['gaussian_kde_copula_gen']
