+++ sklarpy/copulas/_distributions/_student_t.py
@@ -6,7 +6,7 @@ from typing import Union
 
 from sklarpy.copulas._prefit_dists import PreFitCopula
 from sklarpy.copulas._fitted_dists import FittedCopula
-from sklarpy._utils import Params
+from sklarpy.utils._params import Params
 from sklarpy.copulas import MarginalFitter
 
 __all__ = ['student_t_copula_gen']
