+++ sklarpy/multivariate/_distributions/_archimedean.py
@@ -10,7 +10,9 @@ import scipy.optimize
 
 from sklarpy.multivariate._prefit_dists import PreFitContinuousMultivariate
 from sklarpy.multivariate._fitted_dists import FittedContinuousMultivariate
-from sklarpy._utils import get_iterator, FitError, Params
+from sklarpy.utils._params import Params
+from sklarpy.utils._iterator import get_iterator
+from sklarpy.utils._errors import FitError
 from sklarpy.misc import debye
 
 __all__ = ['multivariate_clayton_gen', 'multivariate_gumbel_gen',
