+++ sklarpy/multivariate/_distributions/_gaussian_kde.py
@@ -6,7 +6,7 @@ from typing import Tuple, Union, Callable
 
 from sklarpy.multivariate._prefit_dists import PreFitContinuousMultivariate
 from sklarpy.multivariate._fitted_dists import FittedContinuousMultivariate
-from sklarpy._utils import Params
+from sklarpy.utils._params import Params
 
 __all__ = ['multivariate_gaussian_kde_gen']
