+++ sklarpy/utils/_iterator.py
